putting in redactions for MS voluntarily
While Microsot has said repeatedly that THIS IS NOT A VULNERABILITY, was asked to redact some information. Doing so on the assumption that they are fixing these issues.  REMEMBER -- THIS IS NOT A VULNERABILITY PER MICROSOFT. THIS IS ALL PUBLICLY AVAILABLE, PUBLICLY-DERIVED, FAIRLY GATHERED DATA. I AM VOLUNTARILY REDACTING PARTS OF THIS DATA FOR NOW.