replacing with MS redacted version
REMINDER: THIS DATA IS ALL PUBLICLY AVAILABLE, DERIVED FROM PUBLIC SOURCES, WHICH MICROSOFT HAS REPEATEDLY SAID ARE NOT VULNERABILITIES. I'm self-censoring some of the data on their behalf, but am only doing this under the assumption that they ARE indeed now fixing things. Again, I did NOT have to censor this. Any kid with a laptop with curl could do this type of enumeration.